explorer personal cabinet display api and planet investigation refactor

+++ explorer-personal-cabinet-service/src/main/java/org/example/dto/starsystem/SystemWithProgress.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SystemWithProgress {
     private Integer systemId;
-    private Integer progress;
+    private Double progress;
 }
